+++ app/controllers/questions_controller.rb
@@ -12,23 +12,6 @@ class QuestionsController < ApplicationController
     redirect_to @question
   end
 
-  def new
-    @book = Book.find(params[:book_id])
-    @create_quiz_choice_form = CreateQuizChoiceForm.new
-  end
-
-  def create
-    @book = Book.find(params[:book_id])
-    @create_quiz_choice_form = CreateQuizChoiceForm.new(create_question_params
-      .merge(user_id: current_user.id, book_id: @book.id))
-    if @create_quiz_choice_form.save
-      redirect_to @book, success: '問題を作成しました'
-    else
-      flash.now[:danger] = '問題の作成に失敗しました'
-      render :new
-    end
-  end
-
   def show
     @question = Question.find(params[:id])
     @book = @question.book
@@ -44,12 +27,5 @@ class QuestionsController < ApplicationController
   def destroy
     @question = current_user.questions.find(params[:id])
     @question.destroy
-  end
-
-  private
-
-  def create_question_params
-    params.require(:create_quiz_choice_form)
-          .permit(:question_content, :correct_choice, :incorrect_choice_1, :incorrect_choice_2, :incorrect_choice_3, :commentary)
   end
 end
