+++ config/routes.rb
@@ -25,8 +25,9 @@ Rails.application.routes.draw do
     collection do
       get :search
     end
-    resources :questions, only: %i[create new show destroy]
+    resources :questions, only: %i[show destroy]
     resources :quiz_descriptions, only: %i[new create]
+    resources :quiz_choices, only: %i[new create]
   end
   resources :likes, only: %i[create destroy]
   resources :choices, only: %i[show]
